fix: WakeAll work correct not only for 32-bit system, but also for 64-bit system

+++ internal/futex/futex_linux.go
@@ -25,5 +25,5 @@ func Wake(addr *uint32) {
 
 func WakeAll(addr *uint32) {
 	syscall.Syscall6(syscall.SYS_FUTEX, uintptr(unsafe.Pointer(addr)),
-		opWake|private, ^uintptr(0)>>1, 0, 0, 0)
+		opWake|private, uintptr(^uint32(0)>>1), 0, 0, 0)
 }
